Change return type of the function "calculateLabelCardinality" of the class IRowWiseLabelMatrix from float64 to float32.

+++ cpp/subprojects/common/include/common/input/label_matrix_c_contiguous.hpp
@@ -117,7 +117,7 @@ class CContiguousLabelMatrix final : public CContiguousConstView<const uint8>, v
 
         bool isSparse() const override;
 
-        float64 calculateLabelCardinality() const override;
+        float32 calculateLabelCardinality() const override;
 
         std::unique_ptr<LabelVector> createLabelVector(uint32 row) const override;
